feat(desktop): complete refactor of inventory tracking for better reliability
