+++ apps/client/src/app/modules/inventory/content-id-linking-popup/content-id-linking-popup.component.ts
@@ -2,7 +2,6 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { NzModalRef } from 'ng-zorro-antd/modal';
 import { AuthFacade } from '../../../+state/auth.facade';
 import { SettingsService } from '../../settings/settings.service';
-import { InventoryFacade } from '../+state/inventory.facade';
 
 @Component({
   selector: 'app-content-id-linking-popup',
@@ -19,12 +18,12 @@ export class ContentIdLinkingPopupComponent {
   characterEntries$ = this.authFacade.characterEntries$;
 
   constructor(private authFacade: AuthFacade, private settings: SettingsService,
-              private modalRef: NzModalRef, private inventoryFacade: InventoryFacade) {
+              private modalRef: NzModalRef) {
   }
 
   selectCharacter(id: number): void {
     this.authFacade.setContentId(id, this.contentId);
-    this.modalRef.close(true);
+    this.modalRef.close(this.contentId);
   }
 
   addNewCharacter(): void {
@@ -36,8 +35,7 @@ export class ContentIdLinkingPopupComponent {
       ...this.settings.ignoredContentIds,
       this.contentId
     ];
-    this.inventoryFacade.setContentId(this.previousContentId);
-    this.modalRef.close(false);
+    this.modalRef.close(this.previousContentId);
   }
 
 }
