+++ apps/client/src/app/modules/inventory/inventory.module.ts
@@ -1,9 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { StoreModule } from '@ngrx/store';
-import { EffectsModule } from '@ngrx/effects';
-import { initialState as inventoryInitialState, INVENTORY_FEATURE_KEY, inventoryReducer } from './+state/inventory.reducer';
-import { InventoryEffects } from './+state/inventory.effects';
 import { CoreModule } from '../../core/core.module';
 import { ContentIdLinkingPopupComponent } from './content-id-linking-popup/content-id-linking-popup.component';
 import { NzAvatarModule } from 'ng-zorro-antd/avatar';
@@ -22,10 +18,6 @@ import { InventoryPositionComponent } from './inventory-position/inventory-posit
   imports: [
     CommonModule,
     CoreModule,
-    StoreModule.forFeature(INVENTORY_FEATURE_KEY, inventoryReducer, {
-      initialState: inventoryInitialState
-    }),
-    EffectsModule.forFeature([InventoryEffects]),
     NzAvatarModule,
     NzDividerModule,
     NzTagModule,
