+++ apps/client/src/app/modules/player-metrics/probes/items-probe.ts
@@ -3,19 +3,19 @@ import { filter, map, startWith, withLatestFrom } from 'rxjs/operators';
 import { ProbeReport } from '../model/probe-report';
 import { Observable } from 'rxjs';
 import { MetricType } from '../model/metric-type';
-import { PacketCaptureTrackerService } from '../../../core/electron/packet-capture-tracker.service';
 import { ContainerType } from '../../../model/user/inventory/container-type';
 import { IpcService } from '../../../core/electron/ipc.service';
 import { ProbeSource } from '../model/probe-source';
+import { InventoryService } from '../../inventory/inventory.service';
 
 export class ItemsProbe extends PlayerMetricProbe {
 
-  constructor(protected ipc: IpcService, private machina: PacketCaptureTrackerService) {
+  constructor(protected ipc: IpcService, private inventoryService: InventoryService) {
     super(ipc);
   }
 
   getReports(): Observable<ProbeReport> {
-    return this.machina.inventoryEvents$.pipe(
+    return this.inventoryService.inventoryEvents$.pipe(
       filter(patch => patch.containerId <= 10006 && [ContainerType.Currency, ContainerType.RetainerGil].indexOf(patch.containerId) === -1),
       withLatestFrom(this.source$, this.ipc.eventPlay4Packets$.pipe(filter(e => e.params[0] === 2), startWith(null))),
       map(([event, source, eventPlay4]) => {
