+++ apps/client/src/app/pages/inventory-optimizer/optimizations/duplicates.ts
@@ -3,13 +3,13 @@ import { InventoryItem } from '../../../model/user/inventory/inventory-item';
 import { UserInventory } from '../../../model/user/inventory/user-inventory';
 import { ListRow } from '../../../modules/list/model/list-row';
 import { TranslateService } from '@ngx-translate/core';
-import { InventoryFacade } from '../../../modules/inventory/+state/inventory.facade';
 import { LazyDataService } from '../../../core/data/lazy-data.service';
 import { ContainerType } from '../../../model/user/inventory/container-type';
+import { InventoryService } from '../../../modules/inventory/inventory.service';
 
 export class Duplicates extends InventoryOptimizer {
 
-  constructor(private translate: TranslateService, private inventoryFacade: InventoryFacade, private lazyData: LazyDataService) {
+  constructor(private translate: TranslateService, private inventoryFacade: InventoryService, private lazyData: LazyDataService) {
     super();
   }
