+++ package.json
@@ -117,7 +117,7 @@
   },
   "private": true,
   "dependencies": {
-    "@ffxiv-teamcraft/pcap-ffxiv": "^0.2.6",
+    "@ffxiv-teamcraft/pcap-ffxiv": "^0.3.2",
     "buffer-reader": "^0.1.0",
     "electron-is-dev": "^1.1.0",
     "electron-log": "^3.0.6",
